+++ packages/blocks/src/resolveBlockTemplate.ts
@@ -1,7 +1,8 @@
+import { hooks } from '@faustwp/core';
 import find from 'lodash/find.js';
-import { ContentBlock } from './components/WordPressBlocksViewer.js';
-import { WordPressBlock } from './components/WordPressBlocksProvider.js';
 import DefaultBlock from './components/DefaultBlock.js';
+import { WordPressBlock } from './components/WordPressBlocksProvider.js';
+import { ContentBlock } from './components/WordPressBlocksViewer.js';
 
 /**
  * This function contains the main resolve logic for matching a provided contentBlock instance with the list of
@@ -9,17 +10,19 @@ import DefaultBlock from './components/DefaultBlock.js';
  *
  * @param contentBlock An instance of a block as retrieved from the API
  * @param blocks A list of available WordPressBlock components to match with the provided contentBlock
+ * @param fallbackBlock A React component that, when specified, will be used when no blocks were resolved.
  * @returns An instance of the WordPressBlock component that matches the the provided contentBlock or a DefaultBlock if no such match exists.
  */
 export default function resolveBlockTemplate(
   contentBlock: ContentBlock | null,
   blocks: WordPressBlock[],
+  fallbackBlock: WordPressBlock | undefined,
 ): WordPressBlock {
   // eslint-disable-next-line no-underscore-dangle
   const namesToCheck = [contentBlock?.name, contentBlock?.__typename].filter(
     Boolean,
-  ) as string[];
-  const block = find(blocks, (item) => {
+  );
+  let block = find(blocks, (item) => {
     for (let i = 0; i < namesToCheck.length; i += 1) {
       if (item?.displayName === namesToCheck[i]) {
         return true;
@@ -33,8 +36,26 @@ export default function resolveBlockTemplate(
     }
     return false;
   });
+
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+  block = hooks.applyFilters('faustBlocksResolveBlockTemplate', block, {
+    contentBlock,
+    blocks,
+    namesToCheck,
+  }) as WordPressBlock | undefined;
+
   if (block) {
     return block;
   }
-  return DefaultBlock as WordPressBlock;
+
+  if (fallbackBlock) {
+    return fallbackBlock;
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+  return hooks.applyFilters(
+    'faustBlocksFallbackBlock',
+    DefaultBlock,
+    {},
+  ) as WordPressBlock;
 }