+++ packages/faustwp-cli/src/stdout/index.ts
@@ -1,0 +1,2 @@
+export { debugLog, errorLog, warnLog, infoLog } from './log.js';
+export { styles } from './styles.js';