+++ packages/faustwp-cli/utils/getWpSecret.ts
@@ -1,3 +1,0 @@
-export function getWpSecret() {
-  return process.env.FAUST_SECRET_KEY || process.env.FAUSTWP_SECRET_KEY;
-}