+++ packages/faustwp-core/src/getProps.ts
@@ -16,7 +16,9 @@ export interface GetNextServerSidePropsConfig<Props = Record<string, unknown>> {
     query?: DocumentNode;
     variables?: (
       context: GetStaticPropsContext | GetServerSidePropsContext,
-    ) => { [key: string]: any };
+    ) => {
+      [key: string]: any;
+    };
   };
   props?: Props;
   notFound?: true;
@@ -26,6 +28,16 @@ export interface GetNextServerSidePropsConfig<Props = Record<string, unknown>> {
 export interface GetNextStaticPropsConfig<Props = Record<string, unknown>>
   extends GetNextServerSidePropsConfig<Props> {
   revalidate?: number | boolean;
+}
+
+export interface FaustPage<Data, Props = void>
+  extends React.FC<
+    { data?: Data; __PAGE_VARIABLES__?: { [key: string]: any } } & Props
+  > {
+  query?: DocumentNode;
+  variables?: (context: GetStaticPropsContext | GetServerSidePropsContext) => {
+    [key: string]: any;
+  };
 }
 
 /**
@@ -52,14 +64,27 @@ export async function getNextStaticProps<Props>(
     };
   }
 
+  const pageVariables = Page?.variables ? Page?.variables(context) : undefined;
+
+  let pageQueryRes;
   if (Page.query) {
-    await apolloClient.query({
+    pageQueryRes = await apolloClient.query({
       query: Page.query,
-      variables: Page?.variables ? Page?.variables(context) : undefined,
+      variables: pageVariables,
     });
   }
 
-  const pageProps = addApolloState(apolloClient, { props: { ...props } });
+  let returnedProps = { ...props };
+
+  if (pageQueryRes?.data) {
+    returnedProps = { ...returnedProps, data: pageQueryRes.data };
+  }
+
+  if (pageVariables) {
+    returnedProps = { ...returnedProps, __PAGE_VARIABLES__: pageVariables };
+  }
+
+  const pageProps = addApolloState(apolloClient, { props: returnedProps });
   pageProps.revalidate = revalidate ?? DEFAULT_ISR_REVALIDATE;
   return pageProps as GetStaticPropsResult<Props>;
 }
@@ -92,15 +117,27 @@ export async function getNextServerSideProps<Props>(
     };
   }
 
+  const pageVariables = Page?.variables ? Page?.variables(context) : undefined;
+
+  let pageQueryRes;
   if (Page.query) {
-    await apolloClient.query({
+    pageQueryRes = await apolloClient.query({
       query: Page.query,
-      variables: Page?.variables ? Page?.variables(context) : undefined,
+      variables: pageVariables,
     });
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+  let returnedProps = { ...props };
+
+  if (pageQueryRes?.data) {
+    returnedProps = { ...returnedProps, data: pageQueryRes.data };
+  }
+
+  if (pageVariables) {
+    returnedProps = { ...returnedProps, __PAGE_VARIABLES__: pageVariables };
+  }
+
   return addApolloState(apolloClient, {
-    props: { ...props },
+    props: returnedProps,
   }) as GetServerSidePropsResult<Props>;
 }