+++ packages/faustwp-cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @faustwp/cli
 
+## 0.2.10
+
+### Patch Changes
+
+- c46eac9: Give human readable errors when `faust generatePossibleTypes` fails. Typically due to having "Public Introspection" disabled
+- 0ad4567: - Improved error handling for scenarios when [WPGraphQL](https://wordpress.org/plugins/wp-graphql/) is unavailable due to it being deactivated or the WordPress site is unavailable.
+
 ## 0.2.7
 
 ### Patch Changes