+++ plugins/faustwp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Faust
 
+## 0.8.5
+
+### Patch Changes
+
+- eaa5e48: Added the `shouldShowFaustToolbar` field on the `viewer` WPGraphQL type to determine if the Faust toolbar should be shown based on user preferences.
+
+## 0.8.4
+
+### Patch Changes
+
+- 43205e1: Bug Fix: "Post and Category URL rewrites" setting ignores protocol of configured front-end site URL
+
 ## 0.8.3
 
 ### Patch Changes